Nope all the way out of autochdir
This just did not work for me. I have found the following to be easier:

`:Tex` will open a new tab with the window explorer in the directory of
the file I am currently viewing. From there I can select the file I want
to open and press `Enter` to open it.