Add an `ftplugin` directory for specific syntaxes
File specific syntax settings were created under `.vim/ftplugin`
(organised in directories based on the language/format, with files
containing specific configuration) to reduce the size of the `.vimrc`
folder.

The syntaxes for JSON and Python were ported to the `ftplugin` folder,
and a new syntax declaration for Markdown was added, which removes the
line numbers and adds spell checking.