Update the install script files
The initial problem with the installation script was trying to create
symbolic links for folders that contained files different to the linked
folder.

Instead, a more specific approach was taken by adding more dotfiles that
can be linked, excluding the dotfiles that cannot (e.g. Vim plugin
files). Things such as themes and autoload files were added to both the
install scripts and repository.