Replace `vim-plug` with native plugin management
Rather than using `vim-plug`, Vim 8's native plugin management is used
in conjuction with Git submodules so that plugins can be manually
managed (and version upgrades are managed in the Git repository as
well).

Also removed some plugins that were not being used. They can be readded
as optional plugins in the future if desired.