Restructure and add Git aliases
The aliases were organised a bit to be in a somewhat alphabetical order
(for the short commands, at least). Commands that are nouns have been
grouped by theme.

The `git lost` command was removed due to the pun not helping me
remember what on Earth it actually did, and `git nope` was renamed to
the more appropriate `git scrap` to better represent what it does.

`size`, `where` and various rebase aliases were added to shorten lengthy
commands.